// frontend/src/app/api/model/apiKey.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiKey { 
    /**
     * The unique identifier of this api key
     */
    readonly id: number;
    /**
     * The login of the user link to the api key
     */
    login: string;
}