// frontend/src/app/api/model/role.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Role = 'user' | 'admin:read' | 'admin:write' | 'admin:prod' | 'treasurer:read' | 'treasurer:write' | 'network:read' | 'network:write' | 'network:write:prod' | 'network:write:dev' | 'network:write:hosting';

export const Role = {
    User: 'user' as Role,
    AdminRead: 'admin:read' as Role,
    AdminWrite: 'admin:write' as Role,
    AdminProd: 'admin:prod' as Role,
    TreasurerRead: 'treasurer:read' as Role,
    TreasurerWrite: 'treasurer:write' as Role,
    NetworkRead: 'network:read' as Role,
    NetworkWrite: 'network:write' as Role,
    NetworkWriteProd: 'network:write:prod' as Role,
    NetworkWriteDev: 'network:write:dev' as Role,
    NetworkWriteHosting: 'network:write:hosting' as Role
};